+++ R/volume_methods.R
@@ -14,7 +14,7 @@ read_cnt.ConnectorDatabricksVolume <- function(connector_object, name, ...) {
     unlink(name)
   })
 
-  return(content)
+  content
 }
 
 #' @description
@@ -44,7 +44,8 @@ write_cnt.ConnectorDatabricksVolume <- function(
     file = tmp_name,
     overwrite = overwrite
   )
-  return(invisible(connector_object))
+
+  invisible(connector_object)
 }
 
 #' @description
@@ -64,7 +65,8 @@ list_content_cnt.ConnectorDatabricksVolume <- function(connector_object, ...) {
     purrr::map(list_of_items, ~ purrr::map(.x, "name")),
     use.names = FALSE
   )
-  return(content_names)
+
+  content_names
 }
 
 #' @description
@@ -76,11 +78,10 @@ list_content_cnt.ConnectorDatabricksVolume <- function(connector_object, ...) {
 #' @return [ConnectorDatabricksVolume] object
 #' @export
 remove_cnt.ConnectorDatabricksVolume <- function(connector_object, name, ...) {
-
   file_path <- file.path(connector_object$full_path, name)
   brickster::db_volume_delete(path = file_path)
 
-  return(invisible(connector_object))
+  invisible(connector_object)
 }
 
 #' @description
@@ -100,7 +101,8 @@ download_cnt.ConnectorDatabricksVolume <- function(
     ...) {
   file_path <- file.path(connector_object$full_path, name)
   brickster::db_volume_read(path = file_path, destination = file, ...)
-  return(invisible(connector_object))
+
+  invisible(connector_object)
 }
 
 #' @description
@@ -109,7 +111,6 @@ download_cnt.ConnectorDatabricksVolume <- function(
 #'  `schema` as defined in when initializing the connector.
 #'
 #' @rdname upload_cnt
-#' @param overwrite Overwrite file if it already exists
 #' @param ... [ConnectorDatabricksVolume]: Additional parameters to pass to the [brickster::db_volume_write()] method
 #'
 #' @return [ConnectorDatabricksVolume] object
@@ -127,7 +128,8 @@ upload_cnt.ConnectorDatabricksVolume <- function(
     overwrite = overwrite,
     ...
   )
-  return(invisible(connector_object))
+
+  invisible(connector_object)
 }
 
 #' Create a directory
@@ -141,15 +143,16 @@ upload_cnt.ConnectorDatabricksVolume <- function(
 create_directory_cnt.ConnectorDatabricksVolume <- function(
     connector_object,
     name,
-    ...,
-    open = TRUE) {
+    open = TRUE,
+    ...) {
   directory_path <- file.path(connector_object$full_path, name)
 
   brickster::db_volume_dir_create(path = directory_path, ...)
   if (open) {
     connector_object <- connector_databricks_volume(full_path = directory_path)
   }
-  return(invisible(connector_object))
+
+  invisible(connector_object)
 }
 
 #' Remove a directory
@@ -165,7 +168,8 @@ remove_directory_cnt.ConnectorDatabricksVolume <- function(
     ...) {
   directory_path <- file.path(connector_object$full_path, name)
   brickster::db_volume_dir_delete(path = directory_path)
-  return(invisible(connector_object))
+
+  invisible(connector_object)
 }
 
 #' @description